+++ webpack.config.js
@@ -2,8 +2,8 @@ const webpack = require('webpack');
 const path = require('path');
 
 module.exports = {
-  context: __dirname + '/client',
-  entry: './index.js',
+  context: path.join(__dirname, '/client'),
+  entry: './index.jsx',
   module: {
     rules: [
       {
@@ -11,13 +11,13 @@ module.exports = {
         exclude: /node_modules/,
         loader: 'babel-loader',
         query: {
-          presets: ['react', 'es2015', 'env']
+          presets: ['react', 'es2015', 'env'],
         },
       },
     ],
   },
   output: {
-    path: __dirname + '/public',
+    path: path.join(__dirname, '/public'),
     filename: 'app.js',
-  }
+  },
 };
